requirements.txt
requests==2.31.0
DateTime==5.4
pandas==2.2.2
spotipy==2.23.0
RapidFuzz==3.12.1
Unidecode==1.3.8
anyio==4.8.0
googletrans==3.1.0a0
pytz==2023.3.post1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
ftfy==6.3.1
streamlit
plotly